+++ app/Modules/Layouts/Views/layouts/_head.blade.php
@@ -1,3 +1,4 @@
+
 <link href="{{ asset('favicon.png') }}" rel="icon" type="image/png">
 {{--make sure to use from the BUILD directory...--}}
 {!! Html::style('plugins/jquery-datetimepicker/jquery.datetimepicker.min.css') !!}
@@ -10,5 +11,5 @@
 {!! Html::script('plugins/jquery-datetimepicker/jquery.datetimepicker.full.min.js') !!}
 <script src="{{ asset('plugins/datatables.net/js/jquery.dataTables.min.js') }}"></script>
 <script src="{{ asset('plugins/datatables.net-bs4/js/dataTables.bootstrap4.min.js') }}"></script>
-<script src='{{ asset('plugins/sweetalert2/sweetalert2.all.js') }}'></script>
+<script src='{{ asset('plugins/sweetalert2/sweetalert2.all.min.js') }}'></script>
 <script src="{{ asset('plugins/autosize/jquery.autosize.min.js') }}" type="text/javascript"></script>