+++ app/Modules/Scheduler/Views/categories/edit.blade.php
@@ -13,7 +13,7 @@
                     <div style="text-align:center" class="buttons">
                         <a class="btn btn-warning float-right" href={!! URL::previous()  !!}><i
                                     class="fa fa-ban"></i> {{ trans('fi.cancel') }} </a>
-                        <button type="submit" class="btn btn-success float-right"><i
+                        <button type="submit" class="btn btn-primary float-right"><i
                                     class="fa fa-save"></i> {{ trans('fi.update') }}</button>
                     </div>
                 </h3>