+++ app/Modules/Scheduler/Views/schedule/recurringEventEdit.blade.php
@@ -12,7 +12,7 @@
                         <h3 class="card-title"><i
                                     class="fa fa-edit fa-fw"></i> {{ trans('fi.'.$title) }}
                                 <a class="btn btn-warning float-right" href={!! URL::previous()  !!}><i class="fa fa-ban"></i> {{ trans('fi.cancel') }} </a>
-                                <button type="submit" class="btn btn-success float-right"><i class="fa fa-save"></i> {{ trans('fi.'.$title) }} </button>
+                                <button type="submit" class="btn btn-primary float-right"><i class="fa fa-save"></i> {{ trans('fi.'.$title) }} </button>
                         </h3></div>
                     <div class="card-body">
                         {!! Form::hidden('id') !!}