+++ app/Modules/Settings/Views/settings/_system.blade.php
@@ -1,35 +1,30 @@
 <div class="row">
+    <div class="col-md-4">
+        <div class="form-group">
+            <label>{{ trans('fi.enabled_modules') }}</label>
+            <div class="col-lg-8 col-sm-8">
+                @foreach (\FI\Modules\Settings\Models\Setting::$modules as $entityType => $value)
+                    <div class="form-check">
+                        <label for="enabledModules{{ $value}}">
+                            <input name="enabledModules[]" id="enabledModules{{ $value}}" type="checkbox"
+                                   {{ (new \FI\Modules\Settings\Models\Setting())->isModuleEnabled($entityType) ? 'checked="checked"' : '' }} value="{{ $value }}"> {{ trans("fi.{$entityType}") }}
+                        </label>
+                    </div>
+                @endforeach
+            </div>
+        </div>
+    </div>
     <div class="form-group col-md-4">
         <label>Application URL: </label>
         {!! Form::text('app_url',config('app.url'), ['class' => 'form-control', 'readonly']) !!}
-    </div>
-</div>
-<div class="row">
-    <div class="form-group col-md-4">
         <label>Debug Enabled? </label>
         {!! Form::select('debug', ['0' => 'No', '1' => 'Yes'], config('app.debug'), ['class' => 'form-control', 'disabled']) !!}
-    </div>
-</div>
-<div class="row">
-    <div class="form-group col-md-4">
         <label>Database Driver: </label>
         {!! Form::text('db_driver',config('database.connections.mysql.driver'), ['class' => 'form-control', 'readonly']) !!}
-    </div>
-</div>
-<div class="row">
-    <div class="form-group col-md-4">
         <label>Database Host: </label>
         {!! Form::text('db_host',env('DB_HOST', 'empty'), ['class' => 'form-control', 'readonly']) !!}
-    </div>
-</div>
-<div class="row">
-    <div class="form-group col-md-4">
         <label>Database: </label>
         {!! Form::text('db_database',env('DB_DATABASE','empty'), ['class' => 'form-control', 'readonly']) !!}
-    </div>
-</div>
-<div class="row">
-    <div class="form-group col-md-4">
         <label>Database UserName: </label>
         {!! Form::text('db_username',env('DB_USERNAME', 'empty'), ['class' => 'form-control', 'readonly']) !!}
     </div>