+++ database/migrations/2018_11_08_000100_version_410.php
@@ -17,6 +17,7 @@ class Version410 extends Migration
         deleteTempFiles();
         deleteViewCache();
 
+        Setting::saveByKey('enabledModules', '63');
         Setting::saveByKey('skin','{"headBackground":"purple","headClass":"dark","sidebarBackground":"white","sidebarClass":"light"}');
         Setting::saveByKey('version', '4.1.0');